simplifying headings/heading component
